🐛 fix: close every finding from the fix-program adversarial review
Four independent reviewers over the 12-commit fix program surfaced 34
findings; every real one is closed. The two biggest: the election
scenario could not build at all (its root-level data block never reached
the agent builder, which is constructed with cfg.agents — root data and
fixed_action_sets are now threaded in as reserved builder params, and
silisocs-config-dry-run goes 40/44 -> 44/44 after also migrating the
echo-chamber replications off two retired config shapes) and a reply to
a hallucinated post id escaped as sqlite3.IntegrityError, killing the
agent's whole turn (the engine now validates reply targets like its
sibling actions). Also fixed: r/-prefixed subreddit names break setup;
a de-indented commit escaping the connection pool's cleanup; explicit
null observation/memory_history crashing construction; argument-coercion
failures mislabeled as dropped actions (new counter); validators
false-positiving on custom builders, root fixed_action_sets, disabled
probe stubs, and null sim_role_name; ConfigValidationError swallowing
internal validator bugs; invalid views 500ing three Studio surfaces
(now 422 + a logged nav skip); compose crashing on scalar YAML; panel
409s destroying awaiting placeholders; 500 bodies leaking host paths;
jinja2 missing from the analysis extra; the sample study building 500
agents while declaring 50; demo captures exiting 0 for missing binaries
(bash -lc profile trap); the fix-only hook silently deleting side-effect
imports; and six auto-fixable frozen lint codes drained outright. Docs:
upgrading.md gains seven verified entries, the kernel list and traversal
claims match the enforced contracts, and the import-linter layers now
cover runtime.{model_fields,plugins,provenance,prompts}.

+++ src/silisocs/environments/backends/public_goods/app.py
@@ -160,7 +160,7 @@ class PublicGoodsApp(SimultaneousRoundGame):
                 f"{int(self.endowment)}; got {amount!r}.",
                 committed=False,
             )
-        contribution = int(round(requested))
+        contribution = round(requested)
         error = self.record_choice(agent_name, float(contribution))
         if error:
             return ActionResult(error, committed=False)
